+++ LightBlue.WebService/WebHostSettings.cs
@@ -1,0 +1,13 @@
+﻿namespace LightBlue.WebService
+{
+    public class WebHostSettings
+    {
+        public string Assembly { get; set; }
+        public string Port { get; set; }
+        public string RoleName { get; set; }
+        public string ServiceTitle { get; set; }
+        public string Configuration { get; set; }
+        public bool UseSSL { get; set; }
+        public string Host { get; set; }
+    }
+}